pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "okxsocket"
version = "1.0.0"
description = "OKX cryptocurrency quotes into Telegram bot"
authors = [{name = "Alexander Yusupoff", email = "yusupov.ag@gmail.com"}]
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.10"
keywords = ["okx", "telegram", "crypto", "cryptocurrency", "btc", "eth"]
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License", "Programming Language :: Python", "Programming Language :: Python :: 3"]
dependencies = ["websockets>=10.4", "asyncio>=3.4.3", "python-telegram-bot>=20.0"]

